codespell fix

+++ test/CrossChain/CrossChainOPTellerWithMultiAssetSupport.t.sol
@@ -88,12 +88,12 @@ contract CrossChainOPTellerWithMultiAssetSupportTest is CrossChainBaseTest {
         assertEq(boringVault.balanceOf(address(this)), balBefore - sharesToBridge, "Should have burned shares.");
     }
 
-    function testUniqueIDs() public virtual{
+    function testUniqueIDs() public virtual {
         CrossChainOPTellerWithMultiAssetSupport sourceTeller = CrossChainOPTellerWithMultiAssetSupport(sourceTellerAddr);
         CrossChainOPTellerWithMultiAssetSupport destinationTeller =
             CrossChainOPTellerWithMultiAssetSupport(destinationTellerAddr);
 
-        uint sharesToBridge = 12;
+        uint256 sharesToBridge = 12;
 
         // Bridge shares.
         address to = vm.addr(1);
@@ -111,7 +111,7 @@ contract CrossChainOPTellerWithMultiAssetSupportTest is CrossChainBaseTest {
         uint256 balBefore = boringVault.balanceOf(address(this));
         bytes32 id1 = sourceTeller.bridge{ value: quote }(sharesToBridge, data);
 
-        // preform the exact same bridge again and assert the ids are not the same
+        // perform the exact same bridge again and assert the ids are not the same
         bytes32 id2 = sourceTeller.bridge{ value: quote }(sharesToBridge, data);
 
         assertNotEq(id1, id2, "Id's must be unique");
